requirements.txt
-i https://pypi.org/simple
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
et-xmlfile==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
lxml==4.3.0
more-itertools==5.0.0
numpy==1.15.4
openpyxl==2.5.12
pandas==0.23.4
parsel==1.5.1
pdfminer3k==1.3.1
pillow==5.4.1
pluggy==0.8.0
ply==3.11
py==1.7.0
pyasn1-modules==0.2.3
pyasn1==0.4.5
pycparser==2.19
pydispatcher==2.0.5
pyhamcrest==1.9.0
pyocr==0.5.3
pyopenssl==18.0.0
pytest==4.1.0
python-dateutil==2.7.5
pytz==2018.9
qqai==0.2.0
queuelib==1.5.0
requests==2.21.0
scrapy==1.5.1
service-identity==18.1.0
six==1.12.0
twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
wand==0.5.0
xlrd==1.2.0
zope.interface==4.6.0
